domains/sharedServices/eventProcessing/functions/requirements.txt
azure-functions
azure-eventhub
azure-cosmos
